+++ src/core/skosmos.py
@@ -176,6 +176,10 @@ def get_roles():
     return fetch_data('http://base.uni-ak.ac.at/portfolio/vocabulary/role')
 
 
+def get_statuses():
+    return fetch_data('http://base.uni-ak.ac.at/vocabulary/collection_portfolio_project_status')
+
+
 def get_software_licenses():
     return fetch_data('http://base.uni-ak.ac.at/portfolio/licenses/collection_software_licenses')